+++ src/Clash/Cores/Ethernet/Icmp.hs
@@ -16,7 +16,7 @@ import Clash.Cores.Ethernet.Icmp.IcmpTypes ( IcmpHeader(..), IcmpHeaderLite(..) 
 import Clash.Cores.Ethernet.IP.InternetChecksum ( onesComplementAdd )
 import Clash.Cores.Ethernet.IP.IPv4Types ( IPv4Address, IPv4HeaderLite(..) )
 import Clash.Prelude
-import Protocols ( Circuit, fromSignals, (|>) )
+import Protocols ( Circuit, (|>) )
 import Protocols.Extra.PacketStream
 import Protocols.Extra.PacketStream.Packetizers ( depacketizerC, packetizerC )
